add security around projects

+++ myoo-me/src/main/java/com/myoo/api/resource/AchievementInstanceResource.java
@@ -12,6 +12,7 @@ import javax.ws.rs.core.MediaType;
 
 import com.myoo.api.dao.AchievementDao;
 import com.myoo.api.domain.Achievement;
+import com.myoo.api.service.SecurityService;
 
 @Consumes({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
 @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
@@ -20,6 +21,9 @@ public class AchievementInstanceResource {
 	@Inject
 	private AchievementDao achievementDao;
 
+	@Inject
+	private SecurityService userAccessService;
+
 	@GET
 	public Achievement read(@PathParam("achievementId") String achievementId) {
 		return achievementDao.get(achievementId);
@@ -27,13 +31,21 @@ public class AchievementInstanceResource {
 
 	@POST
 	public Achievement update(@PathParam("achievementId") String achievementId, @Valid Achievement achievement) {
-		achievement.setId(achievementId);
-		return achievementDao.update(achievement);
+		if (userAccessService.isUserAllowed(achievementDao.get(achievementId).getProjectId())) {
+			achievement.setId(achievementId);
+			return achievementDao.update(achievement);
+		} else {
+			throw new SecurityException("User is not allowed to modify Achievements for that Project");
+		}
 	}
 
 	@DELETE
 	public Achievement delete(@PathParam("achievementId") String achievementId) {
-		return achievementDao.delete(achievementId);
+		if (userAccessService.isUserAllowed(achievementDao.get(achievementId).getProjectId())) {
+			return achievementDao.delete(achievementId);
+		} else {
+			throw new SecurityException("User is not allowed to delete Achievements for that Project");
+		}
 	}
 
 }
