+++ myoo-me/src/main/java/com/myoo/api/resource/RecordInstanceResource.java
@@ -12,6 +12,7 @@ import javax.ws.rs.core.MediaType;
 
 import com.myoo.api.dao.RecordDao;
 import com.myoo.api.domain.Record;
+import com.myoo.api.service.SecurityService;
 
 @Consumes({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
 @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
@@ -20,6 +21,9 @@ public class RecordInstanceResource {
 	@Inject
 	private RecordDao recordDao;
 
+	@Inject
+	private SecurityService userAccessService;
+
 	@GET
 	public Record read(@PathParam("recordId") String recordId) {
 		return recordDao.get(recordId);
@@ -27,12 +31,20 @@ public class RecordInstanceResource {
 
 	@POST
 	public Record update(@PathParam("recordId") String recordId, @Valid Record record) {
-		record.setId(recordId);
-		return recordDao.update(record);
+		if (userAccessService.isSelf(recordDao.get(recordId).getUserId())) {
+			record.setId(recordId);
+			return recordDao.update(record);
+		} else {
+			throw new SecurityException("User cannot update Records for other Users");
+		}
 	}
 
 	@DELETE
 	public Record delete(@PathParam("recordId") String recordId) {
-		return recordDao.delete(recordId);
+		if (userAccessService.isSelf(recordDao.get(recordId).getUserId())) {
+			return recordDao.delete(recordId);
+		} else {
+			throw new SecurityException("User cannot delete Records for other Users");
+		}
 	}
 }
