+++ myoo-me/src/main/java/com/myoo/api/resource/SubscriptionInstanceResource.java
@@ -12,6 +12,7 @@ import javax.ws.rs.core.MediaType;
 
 import com.myoo.api.dao.SubscriptionDao;
 import com.myoo.api.domain.Subscription;
+import com.myoo.api.service.SecurityService;
 
 @Consumes({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
 @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
@@ -20,6 +21,9 @@ public class SubscriptionInstanceResource {
 	@Inject
 	private SubscriptionDao subscriptionDao;
 
+	@Inject
+	private SecurityService userAccessService;
+
 	@GET
 	public Subscription read(@PathParam("subscriptionId") String subscriptionId) {
 		return subscriptionDao.get(subscriptionId);
@@ -27,12 +31,20 @@ public class SubscriptionInstanceResource {
 
 	@POST
 	public Subscription update(@PathParam("subscriptionId") String subscriptionId, @Valid Subscription subscription) {
-		subscription.setId(subscriptionId);
-		return subscriptionDao.update(subscription);
+		if (userAccessService.isSelf(subscriptionDao.get(subscriptionId).getUserId())) {
+			subscription.setId(subscriptionId);
+			return subscriptionDao.update(subscription);
+		} else {
+			throw new SecurityException("User cannot update Subscriptions for other Users");
+		}
 	}
 
 	@DELETE
 	public Subscription delete(@PathParam("subscriptionId") String subscriptionId) {
-		return subscriptionDao.delete(subscriptionId);
+		if (userAccessService.isSelf(subscriptionDao.get(subscriptionId).getUserId())) {
+			return subscriptionDao.delete(subscriptionId);
+		} else {
+			throw new SecurityException("User cannto delete Subscriptions for other Users");
+		}
 	}
 }
